requirements.txt
streamlit<=1.34.0
python-dotenv<=1.0.1
openai<=1.28.0
dspy-ai<=2.4.9